refactor: SSR 서버 코드 파일 분리
